+++ src/Command/VersionCommand.php
@@ -4,10 +4,11 @@ declare(strict_types=1);
 
 namespace Smeghead\PhpVariableHardUsage\Command;
 
-final class VersionCommand extends AbstractCommand
+class VersionCommand extends AbstractCommand
 {
-    public function execute(): void
+    public function execute(): int
     {
         $this->printVersion();
+        return 0;
     }
 }